feat: Add mmap ptracer
`mmap tracer` was added for spidermine logic implementation.

An eventless option called PTRACE_O_MMAPTRACE was added to line 139 of
the file `include/uapi/linux/ptrace.h` and shifted 22 times.
Add eventless option as OR operator to PTRACE_O_MASK.

If the PT_MMAPTRACE flag exists in the current->ptrae at the entry of
the syscall and the called systemcall is not mmap(), return syscall
without calling the syscall_trace_enter() and exit the
__syscall_enter_from_user_work().

If the PT_MMAPTRACE flag exists in the current->ptrae at the end of the
syscall and the called systemcall is not mmap(), exit the
syscall_exit_work() without calling the trace_sys_exit() and
ptrace_report_syscall_exit().